feat: attach firewall information to servers
This is not my proudest commit.

Firewall definitions are quite diverse across providers. There are
caveats for individual providers, too.

I am pretty sure that we can find a compact/concise data definition
for representing such diversity, but not today.

For now, I am proceeding with this dirty implementation. It should do
the job.